+++ Changelog.rst
@@ -17,6 +17,9 @@ version 3.15.2
   (https://github.com/NCAS-CMS/cf-python/issues/661)
 * New keyword parameter to `cf.aggregate`: ``cells``
   (https://github.com/NCAS-CMS/cf-python/issues/452)
+* Allow `cf.Data` and other objects that have data (such as
+  `cf.DimensionCoordinate`) to be used as indices
+  (https://github.com/NCAS-CMS/cf-python/issues/677)
 * Fix bug that caused `cf.Field.regrids` and `cf.Field.regridc` to
   fail when the regrid axes are chunked with more than one chunk
   (https://github.com/NCAS-CMS/cf-python/issues/665)